make frequencyError as int16 like RadioHead

+++ common/fusex/trame.h
@@ -102,7 +102,7 @@ typedef struct {
     uint32_t    timestamp;
     int16_t     rssi;
     int16_t     snr;
-    int32_t     frequencyError;
+    int16_t     frequencyError;
 } fxtm_rxfooter_t;
 
 typedef struct {
